mod model of merchant and http method function

+++ digital_wallet/routers/merchants.py
@@ -21,17 +21,21 @@ SIZE_PER_PAGE = 50
 @router.post("")
 async def create_merchant(
     current_user: Annotated[User, Depends(security.get_current_active_user)],
-    item: CreateMerchant,
+    merchant: CreateMerchant,
     session: Annotated[AsyncSession, Depends(models.get_session)],
 ) -> Merchant:
-    data = item.dict()
-    db_merchant = DBMerchant(**data)
+    db_merchant = DBMerchant.model_validate(merchant)
+    db_merchant.user = current_user
+    db_merchant.first_name = current_user.first_name
+    db_merchant.last_name = current_user.last_name
+    db_merchant.email = current_user.email
+    db_merchant.telephone = current_user.telephone
 
     session.add(db_merchant)
     await session.commit()
     await session.refresh(db_merchant)
 
-    return Merchant.from_orm(db_merchant)
+    return Merchant.model_validate(db_merchant)
 
 
 @router.get("")
@@ -50,11 +54,11 @@ async def get_merchants(
         )
     )
 
-    return MerchantList(
-        merchants=db_merchants,
+    return MerchantList.model_validate(
+        dict(merchants=db_merchants,
         page=page,
         page_count=page_count,
-        size_per_page=SIZE_PER_PAGE,
+        size_per_page=SIZE_PER_PAGE,)
     )
 
 
@@ -66,7 +70,7 @@ async def get_merchant(
     if db_merchant is None:
         raise HTTPException(status_code=404, detail="Item not found")
 
-    return Merchant.from_orm(db_merchant)
+    return Merchant.model_validate(db_merchant)
 
 
 @router.put("/{merchant_id}")
@@ -76,17 +80,21 @@ async def update_merchant(
     merchant: UpdateMerchant,
     session: Annotated[AsyncSession, Depends(models.get_session)],
 ) -> Merchant:
+    data = merchant.model_dump()
     db_merchant = await session.get(DBMerchant, merchant_id)
+
     if db_merchant is None:
         raise HTTPException(status_code=404, detail="Item not found")
-    for key, value in merchant.dict().items():
-        setattr(db_merchant, key, value)
-
+    
+    if db_merchant.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Forbidden")
+    
+    db_merchant.sqlmodel_update(data)
     session.add(db_merchant)
     await session.commit()
     await session.refresh(db_merchant)
 
-    return Merchant.from_orm(db_merchant)
+    return Merchant.model_validate(db_merchant)
 
 
 @router.delete("/{merchant_id}")
@@ -96,8 +104,12 @@ async def delete_merchant(
     session: Annotated[AsyncSession, Depends(models.get_session)],
 ) -> dict:
     db_merchant = await session.get(DBMerchant, merchant_id)
+
     if db_merchant is None:
         raise HTTPException(status_code=404, detail="Item not found")
+    
+    if db_merchant.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Forbidden")
 
     await session.delete(db_merchant)
     await session.commit()
